Created a quick-and-dirty solution for exporting GLTF models in a more efficient, lower-level way. This is required for the big Halo Wars maps.

+++ FinModelUtility/Fin/src/exporter/assimp/indirect/AssimpIndirectExporter.cs
@@ -4,17 +4,21 @@ using System.Linq;
 using Assimp;
 
 using fin.exporter.gltf;
+using fin.exporter.gltf.lowlevel;
 using fin.io;
 using fin.model;
 using fin.util.asserts;
 
 using WrapMode = fin.model.WrapMode;
 
+
 namespace fin.exporter.assimp.indirect {
   using FinBlendMode = fin.model.BlendMode;
 
   public class AssimpIndirectExporter : IExporter {
     // You can bet your ass I'm gonna prefix everything with ass.
+
+    public bool LowLevel { get; set; }
 
     public void Export(IFile outputFile, IModel model) {
       var outputPath = outputFile.FullName;
@@ -37,29 +41,34 @@ namespace fin.exporter.assimp.indirect {
         exportFormatId = exportFormatIds.First();
       }
 
-      var assScene = new Scene();
-      assScene.RootNode = new Node("ROOT");
-
-      var inputFile = outputFile.CloneWithExtension(".glb");
+      var inputFile = !this.LowLevel
+                          ? outputFile.CloneWithExtension(".glb")
+                          : outputFile.CloneWithExtension(".gltf");
       var inputPath = inputFile.FullName;
 
-      var gltfExporter = new GltfExporter();
-      gltfExporter.UvIndices = true;
-      gltfExporter.Embedded = true;
-      gltfExporter.Export(inputFile, model);
+      IGltfExporter gltfExporter = !this.LowLevel
+                                       ? new GltfExporter()
+                                       : new LowLevelGltfExporter();
 
-      var sc = ctx.ImportFile(inputPath);
-      File.Delete(inputPath);
+      Scene? assScene = null;
+      if (!this.LowLevel) {
+        gltfExporter.UvIndices = true;
+        gltfExporter.Embedded = true;
+        gltfExporter.Export(inputFile, model);
 
-      // Importing the pre-generated GLTF file does most of the hard work off
-      // the bat: generating the mesh with properly weighted bones.
+        assScene = ctx.ImportFile(inputPath);
+        File.Delete(inputPath);
 
-      // Bone orientation is already correct, you just need to enable
-      // "Automatic Bone Orientation" if importing in Blender.
+        // Importing the pre-generated GLTF file does most of the hard work off
+        // the bat: generating the mesh with properly weighted bones.
 
-      new AssimpIndirectAnimationFixer().Fix(model, sc);
-      new AssimpIndirectUvFixer().Fix(model, sc);
-      new AssimpIndirectTextureFixer().Fix(model, sc);
+        // Bone orientation is already correct, you just need to enable
+        // "Automatic Bone Orientation" if importing in Blender.
+
+        new AssimpIndirectAnimationFixer().Fix(model, assScene);
+        new AssimpIndirectUvFixer().Fix(model, assScene);
+        new AssimpIndirectTextureFixer().Fix(model, assScene);
+      }
 
       // Reexports the GLTF version in case the FBX version is screwed up.
       gltfExporter.UvIndices = false;
@@ -68,15 +77,17 @@ namespace fin.exporter.assimp.indirect {
           new FinFile(inputFile.FullName.Replace(".glb", "_gltf.glb")),
           model);
 
-      // Finally exports the fbx version.
-      // TODO: Are these all safe to include?
-      var preProcessing =
-          PostProcessSteps.FindInvalidData |
-          PostProcessSteps.JoinIdenticalVertices;
+      if (assScene != null) {
+        // Finally exports the fbx version.
+        // TODO: Are these all safe to include?
+        var preProcessing =
+            PostProcessSteps.FindInvalidData |
+            PostProcessSteps.JoinIdenticalVertices;
 
-      var success =
-          ctx.ExportFile(sc, outputPath, exportFormatId, preProcessing);
-      Asserts.True(success, "Failed to export model.");
+        var success =
+            ctx.ExportFile(assScene, outputPath, exportFormatId, preProcessing);
+        Asserts.True(success, "Failed to export model.");
+      }
     }
   }
 }
